Fix of Travis build failure caused by the last phpstan update
The currently used phpstan version includes a new check for iterable types.
If a function argument is iterable, phpstan checks that value type
of its elements is specified in PHPDoc.

Example - if a function accepts array of integers as a parameter,
the corresponding PHPDoc part should change from "array" to "array<int>".

This commit added compliant PHPDoc to functions with iterable arguments.